chore(source): code cleanup
