+++ source/annotations/processors_test.go
@@ -14,6 +14,7 @@ limitations under the License.
 package annotations
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -147,6 +148,46 @@ func TestTTLFromAnnotations(t *testing.T) {
 			resource:    "test-resource",
 			expectedTTL: endpoint.TTL(999999),
 		},
+		{
+			name:        "TTL annotation not present",
+			annotations: map[string]string{"foo": "bar"},
+			expectedTTL: endpoint.TTL(0),
+		},
+		{
+			name:        "TTL annotation value is not a number",
+			annotations: map[string]string{TtlKey: "foo"},
+			expectedTTL: endpoint.TTL(0),
+		},
+		{
+			name:        "TTL annotation value is empty",
+			annotations: map[string]string{TtlKey: ""},
+			expectedTTL: endpoint.TTL(0),
+		},
+		{
+			name:        "TTL annotation value is negative number",
+			annotations: map[string]string{TtlKey: "-1"},
+			expectedTTL: endpoint.TTL(0),
+		},
+		{
+			name:        "TTL annotation value is too high",
+			annotations: map[string]string{TtlKey: fmt.Sprintf("%d", 1<<32)},
+			expectedTTL: endpoint.TTL(0),
+		},
+		{
+			name:        "TTL annotation value is set correctly using integer",
+			annotations: map[string]string{TtlKey: "60"},
+			expectedTTL: endpoint.TTL(60),
+		},
+		{
+			name:        "TTL annotation value is set correctly using duration (whole)",
+			annotations: map[string]string{TtlKey: "10m"},
+			expectedTTL: endpoint.TTL(600),
+		},
+		{
+			name:        "TTL annotation value is set correctly using duration (fractional)",
+			annotations: map[string]string{TtlKey: "20.5s"},
+			expectedTTL: endpoint.TTL(20),
+		},
 	}
 
 	for _, tt := range tests {
@@ -183,6 +224,126 @@ func TestGetAliasFromAnnotations(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			result := hasAliasFromAnnotations(tt.annotations)
+			assert.Equal(t, tt.expected, result)
+		})
+	}
+}
+
+func TestHostnamesFromAnnotations(t *testing.T) {
+	tests := []struct {
+		name        string
+		annotations map[string]string
+		expected    []string
+	}{
+		{
+			name:        "no hostname annotation",
+			annotations: map[string]string{},
+			expected:    nil,
+		},
+		{
+			name: "single hostname annotation",
+			annotations: map[string]string{
+				HostnameKey: "example.com",
+			},
+			expected: []string{"example.com"},
+		},
+		{
+			name: "multiple hostname annotations",
+			annotations: map[string]string{
+				HostnameKey: "example.com,example.org",
+			},
+			expected: []string{"example.com", "example.org"},
+		},
+		{
+			name: "hostname annotation with spaces",
+			annotations: map[string]string{
+				HostnameKey: " example.com , example.org ",
+			},
+			expected: []string{"example.com", "example.org"},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := HostnamesFromAnnotations(tt.annotations)
+			assert.Equal(t, tt.expected, result)
+		})
+	}
+}
+
+func TestSplitHostnameAnnotation(t *testing.T) {
+	tests := []struct {
+		name       string
+		annotation string
+		expected   []string
+	}{
+		{
+			name:       "empty annotation",
+			annotation: "",
+			expected:   []string{""},
+		},
+		{
+			name:       "single hostname",
+			annotation: "example.com",
+			expected:   []string{"example.com"},
+		},
+		{
+			name:       "multiple hostnames",
+			annotation: "example.com,example.org",
+			expected:   []string{"example.com", "example.org"},
+		},
+		{
+			name:       "hostnames with spaces",
+			annotation: " example.com , example.org ",
+			expected:   []string{"example.com", "example.org"},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := SplitHostnameAnnotation(tt.annotation)
+			assert.Equal(t, tt.expected, result)
+		})
+	}
+}
+
+func TestInternalHostnamesFromAnnotations(t *testing.T) {
+	tests := []struct {
+		name        string
+		annotations map[string]string
+		expected    []string
+	}{
+		{
+			name:        "no internal hostname annotation",
+			annotations: map[string]string{},
+			expected:    nil,
+		},
+		{
+			name: "single internal hostname annotation",
+			annotations: map[string]string{
+				InternalHostnameKey: "internal.example.com",
+			},
+			expected: []string{"internal.example.com"},
+		},
+		{
+			name: "multiple internal hostname annotations",
+			annotations: map[string]string{
+				InternalHostnameKey: "internal.example.com,internal.example.org",
+			},
+			expected: []string{"internal.example.com", "internal.example.org"},
+		},
+		{
+			name: "internal hostname annotation with spaces",
+			annotations: map[string]string{
+				InternalHostnameKey: " internal.example.com , internal.example.org ",
+			},
+			expected: []string{"internal.example.com", "internal.example.org"},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := InternalHostnamesFromAnnotations(tt.annotations)
 			assert.Equal(t, tt.expected, result)
 		})
 	}
