+++ source/source_test.go
@@ -17,285 +17,143 @@ limitations under the License.
 package source
 
 import (
-	"fmt"
-	"strconv"
+	"context"
+	"reflect"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
-
-	"sigs.k8s.io/external-dns/endpoint"
 )
 
-func TestGetTTLFromAnnotations(t *testing.T) {
-	for _, tc := range []struct {
-		title       string
-		annotations map[string]string
-		expectedTTL endpoint.TTL
+func TestParseTemplate(t *testing.T) {
+	for _, tt := range []struct {
+		name                     string
+		annotationFilter         string
+		fqdnTemplate             string
+		combineFQDNAndAnnotation bool
+		expectError              bool
 	}{
 		{
-			title:       "TTL annotation not present",
-			annotations: map[string]string{"foo": "bar"},
-			expectedTTL: endpoint.TTL(0),
+			name:         "invalid template",
+			expectError:  true,
+			fqdnTemplate: "{{.Name",
 		},
 		{
-			title:       "TTL annotation value is not a number",
-			annotations: map[string]string{ttlAnnotationKey: "foo"},
-			expectedTTL: endpoint.TTL(0),
+			name:        "valid empty template",
+			expectError: false,
 		},
 		{
-			title:       "TTL annotation value is empty",
-			annotations: map[string]string{ttlAnnotationKey: ""},
-			expectedTTL: endpoint.TTL(0),
+			name:         "valid template",
+			expectError:  false,
+			fqdnTemplate: "{{.Name}}-{{.Namespace}}.ext-dns.test.com",
 		},
 		{
-			title:       "TTL annotation value is negative number",
-			annotations: map[string]string{ttlAnnotationKey: "-1"},
-			expectedTTL: endpoint.TTL(0),
+			name:         "valid template",
+			expectError:  false,
+			fqdnTemplate: "{{.Name}}-{{.Namespace}}.ext-dns.test.com, {{.Name}}-{{.Namespace}}.ext-dna.test.com",
 		},
 		{
-			title:       "TTL annotation value is too high",
-			annotations: map[string]string{ttlAnnotationKey: fmt.Sprintf("%d", 1<<32)},
-			expectedTTL: endpoint.TTL(0),
+			name:                     "valid template",
+			expectError:              false,
+			fqdnTemplate:             "{{.Name}}-{{.Namespace}}.ext-dns.test.com, {{.Name}}-{{.Namespace}}.ext-dna.test.com",
+			combineFQDNAndAnnotation: true,
 		},
 		{
-			title:       "TTL annotation value is set correctly using integer",
-			annotations: map[string]string{ttlAnnotationKey: "60"},
-			expectedTTL: endpoint.TTL(60),
-		},
-		{
-			title:       "TTL annotation value is set correctly using duration (whole)",
-			annotations: map[string]string{ttlAnnotationKey: "10m"},
-			expectedTTL: endpoint.TTL(600),
-		},
-		{
-			title:       "TTL annotation value is set correctly using duration (fractional)",
-			annotations: map[string]string{ttlAnnotationKey: "20.5s"},
-			expectedTTL: endpoint.TTL(20),
+			name:             "non-empty annotation filter label",
+			expectError:      false,
+			annotationFilter: "kubernetes.io/ingress.class=nginx",
 		},
 	} {
-		t.Run(tc.title, func(t *testing.T) {
-			ttl := getTTLFromAnnotations(tc.annotations, "resource/test")
-			assert.Equal(t, tc.expectedTTL, ttl)
+		t.Run(tt.name, func(t *testing.T) {
+			_, err := parseTemplate(tt.fqdnTemplate)
+			if tt.expectError {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+			}
 		})
 	}
 }
 
-func TestGetProviderSpecificCloudflareAnnotations(t *testing.T) {
-	for _, tc := range []struct {
-		title         string
-		annotations   map[string]string
-		expectedKey   string
-		expectedValue bool
+func TestFqdnTemplate(t *testing.T) {
+	tests := []struct {
+		name          string
+		fqdnTemplate  string
+		expectedError bool
 	}{
 		{
-			title:         "Cloudflare proxied annotation is set correctly to true",
-			annotations:   map[string]string{CloudflareProxiedKey: "true"},
-			expectedKey:   CloudflareProxiedKey,
-			expectedValue: true,
+			name:          "empty template",
+			fqdnTemplate:  "",
+			expectedError: false,
 		},
 		{
-			title:         "Cloudflare proxied annotation is set correctly to false",
-			annotations:   map[string]string{CloudflareProxiedKey: "false"},
-			expectedKey:   CloudflareProxiedKey,
-			expectedValue: false,
+			name:          "valid template",
+			fqdnTemplate:  "{{ .Name }}.example.com",
+			expectedError: false,
 		},
-		{
-			title: "Cloudflare proxied annotation among another annotations is set correctly to true",
-			annotations: map[string]string{
-				"random annotation 1": "random value 1",
-				CloudflareProxiedKey:  "false",
-				"random annotation 2": "random value 2",
-			},
-			expectedKey:   CloudflareProxiedKey,
-			expectedValue: false,
-		},
-	} {
-		t.Run(tc.title, func(t *testing.T) {
-			providerSpecificAnnotations, _ := getProviderSpecificAnnotations(tc.annotations)
-			for _, providerSpecificAnnotation := range providerSpecificAnnotations {
-				if providerSpecificAnnotation.Name == tc.expectedKey {
-					assert.Equal(t, strconv.FormatBool(tc.expectedValue), providerSpecificAnnotation.Value)
-					return
-				}
-			}
-			t.Errorf("Cloudflare provider specific annotation %s is not set correctly to %v", tc.expectedKey, tc.expectedValue)
-		})
 	}
 
-	for _, tc := range []struct {
-		title         string
-		annotations   map[string]string
-		expectedKey   string
-		expectedValue string
-	}{
-		{
-			title:         "Cloudflare custom hostname annotation is set correctly",
-			annotations:   map[string]string{CloudflareCustomHostnameKey: "a.foo.fancybar.com"},
-			expectedKey:   CloudflareCustomHostnameKey,
-			expectedValue: "a.foo.fancybar.com",
-		},
-		{
-			title: "Cloudflare custom hostname annotation among another annotations is set correctly",
-			annotations: map[string]string{
-				"random annotation 1":       "random value 1",
-				CloudflareCustomHostnameKey: "a.foo.fancybar.com",
-				"random annotation 2":       "random value 2"},
-			expectedKey:   CloudflareCustomHostnameKey,
-			expectedValue: "a.foo.fancybar.com",
-		},
-	} {
-		t.Run(tc.title, func(t *testing.T) {
-			providerSpecificAnnotations, _ := getProviderSpecificAnnotations(tc.annotations)
-			for _, providerSpecificAnnotation := range providerSpecificAnnotations {
-				if providerSpecificAnnotation.Name == tc.expectedKey {
-					assert.Equal(t, tc.expectedValue, providerSpecificAnnotation.Value)
-					return
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			tmpl, err := parseTemplate(tt.fqdnTemplate)
+			if tt.expectedError {
+				assert.Error(t, err)
+				assert.Nil(t, tmpl)
+			} else {
+				assert.NoError(t, err)
+				if tt.fqdnTemplate == "" {
+					assert.Nil(t, tmpl)
+				} else {
+					assert.NotNil(t, tmpl)
 				}
 			}
-			t.Errorf("Cloudflare provider specific annotation %s is not set correctly to %s", tc.expectedKey, tc.expectedValue)
 		})
 	}
 }
 
-func TestGetProviderSpecificAliasAnnotations(t *testing.T) {
-	for _, tc := range []struct {
-		title         string
-		annotations   map[string]string
-		expectedKey   string
-		expectedValue bool
-	}{
-		{
-			title:         "alias annotation is set correctly to true",
-			annotations:   map[string]string{aliasAnnotationKey: "true"},
-			expectedKey:   aliasAnnotationKey,
-			expectedValue: true,
-		},
-		{
-			title: "alias annotation among another annotations is set correctly to true",
-			annotations: map[string]string{
-				"random annotation 1": "random value 1",
-				aliasAnnotationKey:    "true",
-				"random annotation 2": "random value 2",
-			},
-			expectedKey:   aliasAnnotationKey,
-			expectedValue: true,
-		},
-	} {
-		t.Run(tc.title, func(t *testing.T) {
-			providerSpecificAnnotations, _ := getProviderSpecificAnnotations(tc.annotations)
-			for _, providerSpecificAnnotation := range providerSpecificAnnotations {
-				if providerSpecificAnnotation.Name == "alias" {
-					assert.Equal(t, strconv.FormatBool(tc.expectedValue), providerSpecificAnnotation.Value)
-					return
-				}
-			}
-			t.Errorf("provider specific annotation alias is not set correctly to %v", tc.expectedValue)
-		})
-	}
-
-	for _, tc := range []struct {
-		title       string
-		annotations map[string]string
-	}{
-		{
-			title:       "alias annotation is set to false",
-			annotations: map[string]string{aliasAnnotationKey: "false"},
-		},
-		{
-			title: "alias annotation is not set",
-			annotations: map[string]string{
-				"random annotation 1": "random value 1",
-				"random annotation 2": "random value 2",
-			},
-		},
-	} {
-		t.Run(tc.title, func(t *testing.T) {
-			providerSpecificAnnotations, _ := getProviderSpecificAnnotations(tc.annotations)
-			for _, providerSpecificAnnotation := range providerSpecificAnnotations {
-				if providerSpecificAnnotation.Name == "alias" {
-					t.Error("provider specific annotation alias is not expected to be set")
-				}
-			}
-
-		})
-	}
+type mockInformerFactory struct {
+	syncResults map[reflect.Type]bool
 }
 
-func TestGetProviderSpecificIdentifierAnnotations(t *testing.T) {
-	for _, tc := range []struct {
-		title              string
-		annotations        map[string]string
-		expectedResult     map[string]string
-		expectedIdentifier string
+func (m *mockInformerFactory) WaitForCacheSync(stopCh <-chan struct{}) map[reflect.Type]bool {
+	return m.syncResults
+}
+
+func TestWaitForCacheSync(t *testing.T) {
+	tests := []struct {
+		name        string
+		syncResults map[reflect.Type]bool
+		expectError bool
 	}{
 		{
-			title: "aws- provider specific annotations are set correctly",
-			annotations: map[string]string{
-				"external-dns.alpha.kubernetes.io/aws-annotation-1": "value 1",
-				SetIdentifierKey: "id1",
-				"external-dns.alpha.kubernetes.io/aws-annotation-2": "value 2",
-			},
-			expectedResult: map[string]string{
-				"aws/annotation-1": "value 1",
-				"aws/annotation-2": "value 2",
-			},
-			expectedIdentifier: "id1",
+			name:        "all caches synced",
+			syncResults: map[reflect.Type]bool{reflect.TypeOf(""): true},
+			expectError: false,
 		},
 		{
-			title: "scw- provider specific annotations are set correctly",
-			annotations: map[string]string{
-				"external-dns.alpha.kubernetes.io/scw-annotation-1": "value 1",
-				SetIdentifierKey: "id1",
-				"external-dns.alpha.kubernetes.io/scw-annotation-2": "value 2",
-			},
-			expectedResult: map[string]string{
-				"scw/annotation-1": "value 1",
-				"scw/annotation-2": "value 2",
-			},
-			expectedIdentifier: "id1",
+			name:        "some caches not synced",
+			syncResults: map[reflect.Type]bool{reflect.TypeOf(""): false},
+			expectError: true,
 		},
 		{
-			title: "ibmcloud- provider specific annotations are set correctly",
-			annotations: map[string]string{
-				"external-dns.alpha.kubernetes.io/ibmcloud-annotation-1": "value 1",
-				SetIdentifierKey: "id1",
-				"external-dns.alpha.kubernetes.io/ibmcloud-annotation-2": "value 2",
-			},
-			expectedResult: map[string]string{
-				"ibmcloud-annotation-1": "value 1",
-				"ibmcloud-annotation-2": "value 2",
-			},
-			expectedIdentifier: "id1",
+			name:        "context timeout",
+			syncResults: map[reflect.Type]bool{reflect.TypeOf(""): false},
+			expectError: true,
 		},
-		{
-			title: "webhook- provider specific annotations are set correctly",
-			annotations: map[string]string{
-				"external-dns.alpha.kubernetes.io/webhook-annotation-1": "value 1",
-				SetIdentifierKey: "id1",
-				"external-dns.alpha.kubernetes.io/webhook-annotation-2": "value 2",
-			},
-			expectedResult: map[string]string{
-				"webhook/annotation-1": "value 1",
-				"webhook/annotation-2": "value 2",
-			},
-			expectedIdentifier: "id1",
-		},
-	} {
-		t.Run(tc.title, func(t *testing.T) {
-			providerSpecificAnnotations, identifier := getProviderSpecificAnnotations(tc.annotations)
-			assert.Equal(t, tc.expectedIdentifier, identifier)
-			for expectedAnnotationKey, expectedAnnotationValue := range tc.expectedResult {
-				expectedResultFound := false
-				for _, providerSpecificAnnotation := range providerSpecificAnnotations {
-					if providerSpecificAnnotation.Name == expectedAnnotationKey {
-						assert.Equal(t, expectedAnnotationValue, providerSpecificAnnotation.Value)
-						expectedResultFound = true
-						break
-					}
-				}
-				if !expectedResultFound {
-					t.Errorf("provider specific annotation %s has not been set", expectedAnnotationKey)
-				}
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			ctx, cancel := context.WithTimeout(context.Background(), 50*time.Millisecond)
+			defer cancel()
+
+			factory := &mockInformerFactory{syncResults: tt.syncResults}
+			err := waitForCacheSync(ctx, factory)
+
+			if tt.expectError {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
 			}
 		})
 	}
